sovd-token-service/requirements.txt
fastapi
uvicorn[standard]
pyjwt
cryptography
python-jose
python-multipart
requests
aiofiles